--- src/main/java/team/marco/vouchermanagementsystem/Voucher.java
@@ -1,4 +1,0 @@
-package team.marco.vouchermanagementsystem;
-
-public interface Voucher {
-}
