+++ src/main/java/team/marco/vouchermanagementsystem/model/Voucher.java
@@ -1,0 +1,4 @@
+package team.marco.vouchermanagementsystem.model;
+
+public interface Voucher {
+}
