+++ src/main/java/team/marco/vouchermanagementsystem/repository/VoucherRepository.java
@@ -1,4 +1,6 @@
-package team.marco.vouchermanagementsystem;
+package team.marco.vouchermanagementsystem.repository;
+
+import team.marco.vouchermanagementsystem.model.Voucher;
 
 import java.util.List;
